Cast MONEY type to NUMERIC in sql retrieval scripts

+++ sql/DatabaseCreationScripts/CreateViews/6.0.2 Get_Tools_View.sql
@@ -1,5 +1,5 @@
 CREATE OR REPLACE VIEW Get_Tools
 AS
-SELECT a.ID, a.acquisition_date, a.value, t.description, a.chapter_id, a.donor_id
+SELECT a.ID, a.acquisition_date, a.value::NUMERIC, t.description, a.chapter_id, a.donor_id
 FROM Asset a
 JOIN Tool t ON a.ID = t.ID
